code cleanup and start tertiary reserve transfer
